service_a/requirements.txt
flask
kafka-python
boto3
python-dotenv
flasgger
